Not functional progress commit
We are working on transitioning to tabs. The change has broken callbacks, but we are getting there. The difficulty is storing the state in conjunction with callbacks - the two don't play well together. We are currently reorganising the state variables to store constructors rather than objects to remake the sections correctly each time. Its proving difficult. We might want to tear a lot down and start again come monday. Good luck